OOP: Update data type and variable
student.cpp s1 -> std1, class.cpp int -> float

+++ OOP/student.cpp
@@ -1,6 +1,6 @@
 class Student {
 public:
-    int rollNumber;
+    float rollNumber;
 private:
     int age;
 public:
